+++ discord/ui/__init__.py
@@ -12,6 +12,7 @@ from .action_row import *
 from .button import *
 from .checkbox import *
 from .checkbox_group import *
+from .constant import *
 from .container import *
 from .core import *
 from .file import *